+add logic.h +moved definitions to the type.h file +lessened the arguments in the allegro function
