+++ logics.h
@@ -1,0 +1,8 @@
+#pragma once
+#include <stdio.h>
+#include "types.h"
+
+int isLegal(struct pos)
+{
+
+}
